+++ src/impl/index.ts
@@ -1,4 +1,5 @@
 export * from './files-to-test-string';
 export * from './read-fake-files';
 export * from './find-file-system-test-case-directories';
-export * from './file-comparison-test-body';
+export * from './run-file-comparison-test-body';
+export * from './file-system-test-case-runs';