+++ src/impl/run-file-comparison-test-body.ts
@@ -3,7 +3,7 @@ import { FilesContainer, TestComparisonStrings } from '../types';
 import { filesToTestString } from './files-to-test-string';
 import { readFakeFiles } from './read-fake-files';
 
-export async function fileComparisonTestBody(
+export async function runFileComparisonTestBody(
   testCasesParentDirectory: string,
   exampleName: string,
   actualFunction: (testCaseDirectory: string) => Promise<FilesContainer>,
@@ -17,15 +17,8 @@ export async function fileComparisonTestBody(
 
   const generatedFiles = await actualFunction(testCaseDirectory);
 
-  return toTestComparisonStrings(testFiles, generatedFiles);
-}
-
-function toTestComparisonStrings(
-  expected: FilesContainer,
-  actual: FilesContainer
-): TestComparisonStrings {
   return {
-    expected: filesToTestString(expected.textFiles, expected.binaryFiles),
-    actual: filesToTestString(actual.textFiles, actual.binaryFiles),
+    expected: filesToTestString(testFiles),
+    actual: filesToTestString(generatedFiles),
   };
 }